requirements.txt
asgiref==3.3.1
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
psycopg2==2.8.6
python-decouple==3.3
dj-static==0.0.6
dj-database-url==0.5.0
